move tracing to test file
