+++ config/config.exs
@@ -9,4 +9,8 @@ config :logger, :console,
   format: "[$level] $metadata\t| $message\n",
   metadata: [:name, :t, :term, :node]
 
+if config_env() == :test do
+  config :craft, :consensus_module, Craft.TracedConsensus
+end
+
 #import_config "#{config_env()}.exs"
