+++ lib/craft/consensus.ex
@@ -119,71 +119,6 @@ defmodule Craft.Consensus do
     :gen_statem.start_link({:local, name(args.name)}, __MODULE__, args, [])
   end
 
-  if Mix.env() == :test do
-    defoverridable start_link: 1
-
-    def start_link(args) do
-      name = args.name
-      args =
-        case Map.pop(args, :consensus_data) do
-          {nil, args} -> args
-          {data, args} -> {data, args}
-        end
-
-      :gen_statem.start_link({:local, name(name)}, Craft.Consensus.Tracer, args, [])
-    end
-
-    defmodule Tracer do
-      @moduledoc """
-      decorator to capture state machine events in test
-
-      :erlang.trace/3 can't guarantee delivery order between trace messages and messages that this process sends, so we decorate instead
-      """
-
-      defdelegate callback_mode, to: Craft.Consensus
-
-      def init({data, args}) when is_struct(data) do
-        # forward logs to testing server
-        remote_group_leader = :rpc.call(node(data.nexus_pid), Process, :whereis, [:init])
-        :logger.update_process_metadata(%{gl: remote_group_leader})
-
-        MemberCache.update(data)
-
-        state = if args[:manual_start], do: :ready_to_test, else: :lonely
-
-        {:ok, state, data}
-      end
-
-      def init(args) do
-        data = %State{
-          State.new(args.name, args.nodes, args.persistence)
-          | nexus_pid: args.nexus_pid,
-            state: :lonely
-        }
-
-        init({data, args})
-      end
-
-      def ready_to_test(:enter, _, _data), do: :keep_state_and_data
-
-      def ready_to_test(:cast, :run, %State{state: state} = data),
-        do: {:next_state, state, data, []}
-
-      def ready_to_test({:call, _from}, :catch_up, _data), do: {:keep_state_and_data, [:postpone]}
-
-      for state <- [:lonely, :receiving_snapshot, :follower, :candidate, :leader] do
-        def unquote(state)(event, msg, data) do
-          send(
-            data.nexus_pid,
-            {DateTime.utc_now(), {:trace, node(), unquote(state), event, msg, data}}
-          )
-
-          apply(Craft.Consensus, unquote(state), [event, msg, data])
-        end
-      end
-    end
-  end
-
   def init(args) do
     Logger.metadata(name: args.name, node: node())
